refactor: content 빼기
